skid_buffer/pyproject.toml
[project]
name = "skid_buffer"
version = "0.1.0"
description = "Skid buffer with axi stream interface"
readme = "README.md"
authors = [
    { name = "Caspar Trittibach", email = "ctrittibach@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "click >= 8.1.7",
    "pytest >= 8.2.1",
    "cocotb >= 1.8.1",
    "cocotb-test >= 0.2.4",
    "cocotbext-axi >= 0.1.24",
]

[tool.pytest.ini_options]
pythonpath = "tb/"

